renaming index files instead of package.js
