+++ client/src/App.js
@@ -6,7 +6,7 @@ import GlobalStyles from "./styles/GlobalStyle";
 import Login from "./pages/Login/Login";
 
 import { getAccessToken } from "./auth";
-import Explorify from "./layout/Explorify";
+import Explorify from "./layout";
 
 const axiosInstance = axios.create({
   baseURL: "https://api.spotify.com/v1",
