+++ client/src/auth/index.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { getHashParams } from "../utils/utils";
+import { getHashParams } from "../utils";
 
 //------------------------------------------------
 //---------> AUTH TOKENS
