+++ client/src/components/AudioFeaturesBar/index.js
