+++ client/src/components/Icon/SpotifyIcon/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Icon from "../Icon";
+import Icon from "..";
 import styled from "styled-components/macro";
 
 const StyledIcon = styled(Icon)`
