--- client/src/components/Icons/Icon/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "Icon.js"
-}
