--- client/src/components/Icons/SpotifyIcon/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "SpotifyIcon.js"
-}
