--- client/src/components/Loader/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "Loader.js"
-}
