+++ client/src/components/PlaceholderImage/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Icon from "../Icons/Icon";
+import Icon from "../Icon";
 import styled from "styled-components/macro";
 
 const PlaceholderWrapper = styled.div`
