--- client/src/components/SongContainer/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "SongContainer.js"
-}
