--- client/src/components/SongItem/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "SongItem.js"
-}
