--- client/src/components/TitleWrapper/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "TitleWrapper.js"
-}
