+++ client/src/container/SearchResults/SearchResults/AlbumResults/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import CardSection from "../../../../components/CardSection/CardSection";
+import CardSection from "../../../../components/CardSection";
 import useSWR from "swr";
 
 const filterArtists = (artists) => {
