--- client/src/layout/Header/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "Header.js"
-}
