+++ client/src/layout/SideNav/AppTitle/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import SpotifyIcon from "../../../components/Icons/SpotifyIcon";
+import SpotifyIcon from "../../../components/Icon/SpotifyIcon";
 import { Link } from "react-router-dom";
 import styled from "styled-components/macro";
 import theme from "../../../styles/theme";
