--- client/src/layout/SideNav/AppTitle/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "AppTitle.js"
-}
