+++ client/src/layout/SideNav/NavItem/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import theme from "../../../styles/theme";
 import mixins from "../../../styles/mixins";
 import { NavLink } from "react-router-dom";
-import Icon from "../../../components/Icons/Icon";
+import Icon from "../../../components/Icon";
 import { capitalizeWord } from "../../../utils";
 
 import styled from "styled-components/macro";
