--- client/src/layout/SideNav/NavItem/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "NavItem.js"
-}
