--- client/src/layout/SideNav/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "SideNav.js"
-}
