--- client/src/layout/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "Explorify.js"
-}
