+++ client/src/pages/Explore/Explore.js
@@ -1,5 +1,5 @@
 import React from "react";
-import SearchContainer from "../../container/SearchResults/SearchContainer";
+import SearchContainer from "../../container/SearchResults";
 import styled from "styled-components/macro";
 
 const FlexContainer = styled.div`
