+++ client/src/pages/Home/Home.js
@@ -1,9 +1,9 @@
 import React from "react";
-import WelcomeUser from "./components/WelcomeUser/WelcomeUser";
-import TopArtist from "../../container/ArtistAndTracks/TopArtist/TopArtist";
-import TopArtistScroller from "../../container/TopArtistScroller/TopArtistScroller";
-import RecentlyPlayedSongs from "../../container/RecentlyPlayedSongs/RecentlyPlayedSongs";
-import TopTracks from "../../container/TopTracks/TopTracks";
+import WelcomeUser from "./components/WelcomeUser";
+import TopArtist from "../../container/ArtistAndTracks/TopArtist";
+import TopArtistScroller from "../../container/TopArtistScroller";
+import RecentlyPlayedSongs from "../../container/RecentlyPlayedSongs";
+import TopTracks from "../../container/TopTracks";
 
 import theme from "../../styles/theme";
 import styled from "styled-components/macro";
